requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
torch==2.0.1
torchvision==0.15.2
timm==0.9.7
Pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
opencv-python==4.8.1.78
reportlab==4.0.4
lime==0.2.0.1
shap==0.42.1
pytorch-grad-cam==1.4.8
scikit-image==0.21.0
